fix: address code review feedback
- Remove FIXED comment from exception handler
- Move multipart 413 test to test_data_extractors.py following project conventions
- Ensure test follows existing test suite patterns
